refactor: extract method to get rid of the duplicated state
